Fix bug in initial ace mode setting for Coffeescript.
Droplet currently allows 'coffeescript' mode to be set using either
'coffee' or 'coffeescript', but the ace mode needs to be 'ace/mode/coffee'.
This change takes normalizes the mode name before setting it in ace;
it also adds a test.  In the future we probably should require just
spelling the mode name as 'coffee'.