Coffeescript: merge inline code definition blocks.
This change avoids treating inline lambda (->) functions as
separate blocks in Coffeescript in two cases:

(1) direct assignment of a lambda

assignable = (x) -> dosomething

Instead of treating the = and the -> as separate blocks, these are
treated as a single block.

(2) use of a lambda as the last argument of a function call

button 'clickme', ->
  dosomething

Instead of treating the function call (button(..)) as a separate
block from the ->, these are treated as a single block.

This is implemented by factoring the handling of a -> code block
into an @addCode method, and separating the socket handling from
the delineation of the block for the ->.  @addCode is called
by the Assign handler when the RHS is a Code node and by the
Call handler with the last arg is a Code node.

Other Code nodes still get their own block, so, for example,
putting the code in parentheses will split it into its own block.