Expansion palette blocks, and a ui test.
This feature allows a palette block to "expand" into another block when
dragging out of the palette.  The idea is that the block rendered in
the palette can be an abbreviation, and the "real" block might be bigger,
for example, it might include longer arguments etc.

To configure an expansion block, do the following in the palette configuration:

      palette: [{
        name: 'Draw',
        color: 'blue',
        blocks: [{
          block: 'pen()',       // this is what you see in the palette
          expansion: 'pen red', // this is what you get when you drag
          title: 'ptest'
        }],

This commit also adds a basic uitest that simulates mouse events
to test the controller, which exercises the expansion feature.