Separate paletteWrapper and paletteElement.
Previously these two controller members referred to the same
element.  Now they are different, one nested in the other.

Now paletteWrapper deals with the sizing of the overall palette pane
but not the clipping or the containing, and paletteElement contains
all the children and does the clipping.  This allows embedders
such as Pencil Code to decorate the palette pane by attaching elements
to paletteWrapper (such as mode- or palette-switching buttons and menus)
without interfering with the working of the palette.